requirements.txt
discord
bs4
python-dotenv
requests
wikipedia
